Move `serde_from` and `SerdeError` to `crate::serde` inline module
